Release v0.3.0

+++ src/error.rs
@@ -1,8 +1,8 @@
 use std::fmt;
 
-#[derive(Debug)]
+#[derive(Clone, Copy, Debug, PartialEq)]
 pub enum Error {
-    OddImgDim((u32, u32)),
+    OddImgDim((u16, u16)),
     EmptyRange,
     NonFinite,
 }
